ca-cd environment variables added
