+++ docs/commands/show.md
@@ -1,0 +1,24 @@
+# terrahub show
+
+```
+Usage: terrahub show [options]
+
+terrahub@0.0.1 (built: 2018-04-07T19:15:39.787Z)
+this command will run `terraform show` across multiple terrahub components
+
+Options:
+  --include, -i 	 List of components to include (comma separated values)
+  --exclude, -x 	 List of components to exclude (comma separated values)
+  --include-regex, -I 	 List of components to include (regex search)
+  --exclude-regex, -X 	 List of components to exclude (regex search)
+  --git-diff, -g 	 List of components to include (git diff)
+  --var, -r 		 Variable(s) to be used by terraform
+  --var-file, -l 	 Variable file(s) to be used by terraform
+  --silent, -s 	 Runs the command silently (without any output)
+  --env, -e 		 Workspace environment
+  --help, -h 		 Show list of available commands
+```
+
+
+## Return
+Back to [all commands](../commands.md)
